+++ 2023/src/main/java/adventofcode/util/AbstractDay.java
@@ -2,7 +2,7 @@ package adventofcode.util;
 
 public abstract class AbstractDay {
 
-    public abstract int solvePart1(String input);
+	public abstract long solvePart1(String input);
 
-    public abstract int solvePart2(String input);
+	public abstract long solvePart2(String input);
 }
